requirements-mcp.txt
# Required for MCP server integration
Pillow
mcp[cli]
